% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{AbsEnergy}
\alias{AbsEnergy}
\title{AbsEnergy}
\usage{
AbsEnergy(arr)
}
\arguments{
\item{arr}{TSA array with the time series.}
}
\value{
TSA array with the Absolute Energy.
}
\description{
Calculates the sum over the square values of the time series.
}
