% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{AbsoluteSumOfChanges}
\alias{AbsoluteSumOfChanges}
\title{AbsoluteSumOfChanges}
\usage{
AbsoluteSumOfChanges(time.series)
}
\arguments{
\item{time.series}{List of arrays of type double containing the time series.}
}
\value{
List with the absoluteSumOfChanges
}
\description{
Calculates the sum over the absolute value of consecutive
changes in the time series
}
