% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{AggregatedLinearTrend}
\alias{AggregatedLinearTrend}
\title{AggregatedLinearTrend}
\usage{
AggregatedLinearTrend(arr, chunk.size, aggregation.function)
}
\arguments{
\item{arr}{TSA array with the time series.}

\item{chunk.size}{The chunk size used to aggregate the data.}

\item{aggregation.function}{Function to be used in the aggregation. It receives an integer which indicates the
function to be applied:
{
  0 : mean,
  1 : median
  2 : min,
  3 : max,
  4 : stdev,
  default : mean
}}
}
\value{
List of TSA Arrays with:
{
    pvalue: The pvalues for all time series.
    rvalue: The rvalues for all time series.
    intercept: The intercept values for all time series.
    slope: The slope for all time series.
    stderrest: The stderr values for all time series.
}
}
\description{
Calculates a linear least-squares regression for values of the time series that were aggregated
over chunks versus the sequence from 0 up to the number of chunks minus one.
}
