\name{airpass}
\alias{airpass}
\docType{data}
\title{Monthly total international airline passengers }
\description{
Monthly total international airline passengers from 01/1960-
12/1971.}
\usage{data(airpass)}
\format{
  The format is:
 Time-Series [1:144] from 1960 to 1972: 112 118 132 129 121 135 148 148 136 119 ...
}
\source{
Box, G. E. P., Jenkins, G. M. and Reinsel, G. C. (1994) Time Series Analysis, Forecasting
and Control. Second Edition. New York: Prentice-Hall.}
\examples{
data(airpass)
## maybe str(airpass) ; plot(airpass) ...
}
\keyword{datasets}
