% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\docType{methods}
\name{asType,Array,character-method}
\alias{asType,Array,character-method}
\title{asType}
\usage{
\S4method{asType}{Array,character}(a, type)
}
\description{
Returns a copy of the TSA array with a the specified base type.
}
