% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{AutoCorrelation}
\alias{AutoCorrelation}
\title{AutoCorrelation}
\usage{
AutoCorrelation(tss, max.lag, unbiased)
}
\arguments{
\item{tss}{List of arrays of type double containing the time series.}

\item{unbiased}{Determines whether it divides by n - lag (if true) or n (if false)}

\item{max_lag}{The maximum lag to compute.}
}
\value{
The autocorrelation value for the given time series.
}
\description{
Calculates the autocorrelation of the specified lag for the given time series.
}
