% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\docType{methods}
\name{bitShiftL,Array,numeric-method}
\alias{bitShiftL,Array,numeric-method}
\title{bitShiftL}
\usage{
\S4method{bitShiftL}{Array,numeric}(e1, n)
}
\description{
Shifts each element of the in put TSA array n positions to left.
}
