% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\docType{methods}
\name{bitShiftR,Array,numeric-method}
\alias{bitShiftR,Array,numeric-method}
\title{bitShifR}
\usage{
\S4method{bitShiftR}{Array,numeric}(e1, n)
}
\description{
Shifts each element of the in put TSA array n positions to right.
}
