% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{C3}
\alias{C3}
\title{C3}
\usage{
C3(tss, lag)
}
\arguments{
\item{lag}{The lag.}

\item{time.series}{List of arrays of type double containing the time series.}
}
\value{
The non-linearity value for the given time series.
}
\description{
Calculates the Schreiber, T. and Schmitz, A. (1997) measure of non-linearity
for the given time series
}
