\name{color}
\alias{color}
\docType{data}
\title{Color property/time series}
\description{ 
Color property from 35 consecutive batches in an industrial process.
}
\usage{data(color)}
\format{
  The format is:
 Time-Series [1:35] from 1 to 35: 67 63 76 66 69 71 72 71 72 72 ...
}
\source{
``The Estimation of Sigma for an X Chart'', Journal of Quality
Technology, Vol. 22, No. 3 (July 1990), by Jonathan D. Cryer and Thomas P. Ryan.
}
\examples{
data(color)
## maybe str(color) ; plot(color) ...
}
\keyword{datasets}
