% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{CrossCorrelation}
\alias{CrossCorrelation}
\title{CrossCorrelation}
\usage{
CrossCorrelation(arr.xss, arr.yss, unbiased)
}
\arguments{
\item{unbiased}{Determines whether it divides by n - lag (if true) or n (if false).}

\item{xss.arr}{TSA array with the time series.}

\item{yss.arr}{TSA array with the time series.}
}
\value{
The cross-correlation value for the given time series.
}
\description{
Calculates the cross-correlation of the given time series.
}
