% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{CwtCoefficients}
\alias{CwtCoefficients}
\title{CwtCoefficients}
\usage{
CwtCoefficients(arr, arr.widths, coeff, w)
}
\arguments{
\item{arr}{TSA array with the time series.}

\item{arr.widths}{Widths. TSA array.}

\item{coeff}{Coefficient of interest.}

\item{w}{Width of interest.}
}
\value{
Result of calculated coefficients.
}
\description{
Calculates a Continuous wavelet transform for the Ricker wavelet, also known as
the "Mexican hat wavelet".
}
