\name{days}
\alias{days}
\docType{data}
\title{Number of days between payment to Winegard Corp. / time series}
\description{
Accounts receivable data. Number of days until a distributor of
Winegard Company products pays their account.
}
\usage{data(days)}
\format{
  The format is:
 Time-Series [1:130] from 1 to 130: 39 39 41 26 28 28 25 26 24 38 ...
}
\source{
Personal
communication from Mark Selergren, Vice President, Winegard,
Inc., Burlington, Iowa.
}
\examples{
data(days)
## maybe str(days) ; plot(days) ...
}
\keyword{datasets}
