% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{DecimalScalingNorm}
\alias{DecimalScalingNorm}
\title{DecimalScalingNorm}
\usage{
DecimalScalingNorm(tss)
}
\arguments{
\item{tss}{TSA array with the time series.}
}
\value{
: TSA array with the same dimensions as tss, whose values (time series in dimension 0) have been
normalized by dividing each number by 10^j, where j is the number of integer digits of the max number in
the time series.
}
\description{
Normalizes the given time series according to its maximum value and adjusts each value within the range (-1, 1).
}
