% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{DecimalScalingNormInPlace}
\alias{DecimalScalingNormInPlace}
\title{DecimalScalingNormInPlace}
\usage{
DecimalScalingNormInPlace(tss)
}
\arguments{
\item{tss}{TSA array with the time series.}
}
\description{
Same as decimal_scaling_norm, but it performs the operation in place, without allocating further memory.
}
