\name{deere1}
\alias{deere1}
\docType{data}
\title{Deviations of an industrial process at Deere & Co. -- Series 1} 
\description{
82 consecutive values for the amount of deviation (in 0.000025 inch
units) from a specified target value in an industrial machining process at Deere & Co.
}
\usage{data(deere1)}
\format{
  The format is:
 Time-Series [1:82] from 1 to 82: 3 0 -1 -4 7 3 7 3 3 -1 ...
}
\source{
Personal communication from William F.
Fulkerson, Deere & Co. Technical Center, Moline, Illinois.
}
\examples{
data(deere1)
## maybe str(deere1) ; plot(deere1) ...
}
\keyword{datasets}
