\name{deere2}
\alias{deere2}
\docType{data}
\title{Deviations of an industrial process at Deere & Co. -- Series 2} 
\description{
102 consecutive values for the deviation (in 0.0000025 inch units) from a specified target value.}
\usage{data(deere2)}
\format{
  The format is:
 Time-Series [1:102] from 1 to 102: -18 -24 -17 -27 -37 -34 -8 14 18 7 ...
}
\source{
Personal communication from William F.
Fulkerson, Deere & Co. Technical Center, Moline, Illinois.
}
\examples{
data(deere2)
## maybe str(deere2) ; plot(deere2) ...
}
\keyword{datasets}
