\name{euph}
\alias{euph}
\docType{data}
\title{
A digitized sound file of a B flat 
played on a euphonium}
\description{
A digitized sound file of about 0.4 seconds of a B flat just below
middle C played on a euphonium by one of the authors (JDC), a
member of the group Tempered Brass.}
\usage{data(euph)}
\format{
  The format is:
 Time-Series [1:1105] from 1 to 1105: 0.244 0.635 0.712 0.608 0.317 ...
}
\examples{
data(euph)
## maybe str(euph) ; plot(euph) ...
}
\keyword{datasets}
