\name{explode.s}
\alias{explode.s}
\docType{data}
\title{ A simulated explosive AR(1) series}
\description{
A simulated AR(1) series with the AR(1) coefficient being 3.}
\usage{data(explode.s)}
\format{
  The format is:
 Time-Series [1:8] from 1 to 8:  0.63  0.64  3.72 12.67 39.57 ...
}
\examples{
data(explode.s)
## maybe str(explode.s) ; plot(explode.s) ...
}
\keyword{datasets}
