% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{FftAggregated}
\alias{FftAggregated}
\title{FftAggregated}
\usage{
FftAggregated(arr)
}
\arguments{
\item{arr}{TSA array with the time series.}
}
\value{
The spectral centroid (mean), variance, skew, and kurtosis of the absolute fourier transform
 spectrum.
}
\description{
Calculates the spectral centroid(mean), variance, skew, and kurtosis of the absolute fourier transform
spectrum.
}
