% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{FindBestNMotifs}
\alias{FindBestNMotifs}
\title{FindBestNMotifs}
\usage{
FindBestNMotifs(profile, index, n)
}
\arguments{
\item{profile}{TSA Array with the matrix profile containing the minimum distance of each
subsequence.}

\item{index}{TSA Array with the matrix profile index containing where each minimum occurs.}

\item{n}{Number of motifs to extract}
}
\value{
A list of TSA Arrays with the motif distance, the motif indices and the subsequence indices.
}
\description{
This function extracts the best N motifs from a previously calculated matrix profile.
}
