% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{FirstLocationOfMaximum}
\alias{FirstLocationOfMaximum}
\title{FirstLocationOfMaximum}
\usage{
FirstLocationOfMaximum(tss)
}
\arguments{
\item{tss}{List of arrays of type double containing the time series.}
}
\value{
The first relative location of the maximum value to the length of the timeseries,
for each timeseries.
}
\description{
Calculates the first relative location of the maximal value for each timeseries.
}
