% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{FirstLocationOfMinimum}
\alias{FirstLocationOfMinimum}
\title{FistLocationOfMinimum}
\usage{
FirstLocationOfMinimum(arr)
}
\arguments{
\item{arr}{TSA array with the time series.}
}
\value{
The first relative location of the minimal value of each series.
}
\description{
Calculates the first location of the minimal value of each time series. The position
is calculated relatively to the length of the series.
}
