\name{flow}
\alias{flow}
\docType{data}
\title{
Monthly River Flow for the Iowa River}
\description{
Flow data (in cubic feet per second) for the Iowa river measured at
Wapello, Iowa for the period 09/1958 - 08/2006.}
\usage{data(flow)}
\source{
\code{http://waterdata.usgs.gov/ia/nwis/sw}}
\examples{
data(flow)
## maybe str(flow) ; plot(flow) ...
}
\keyword{datasets}
