% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{getDims}
\alias{getDims}
\title{getDims}
\usage{
getDims(a)
}
\arguments{
\item{a}{data.frame which columns have same types and same dimensions.}
}
\value{
TSA Array dimensions.
}
\description{
Gets the TSA array dimensions.
}
