% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{GetRType}
\alias{GetRType}
\title{GetRType}
\usage{
GetRType(a)
}
\arguments{
\item{a}{R data.frame}
}
\value{
Type of the elements in the data.frame (All should have the same type).
}
\description{
Gets the R type corresponding to the tsa Array type of a datafra.me.
}
