\name{gold}
\alias{gold}
\docType{data}
\title{Gold Price / time series} 
\description{
 Daily price of gold (in $ per troy ounce) for the 252 trading days of 2005}
\usage{data(gold)}
\format{
  The format is:
 Time-Series [1:252] from 1 to 252: 427 426 426 423 421 ...
}
\source{www.lbma.org.uk/2005dailygold.htm}
\examples{
data(gold)
## maybe str(gold) ; plot(gold) ...
}
\keyword{datasets}
