\name{google}
\alias{google}
\docType{data}
\title{
Daily returns of the google stock}
\description{
Daily returns of the google stock from 08/20/04 - 09/13/06.
}
\usage{data(google)}
\format{
  The format is:
 Time-Series [1:521] from 1 to 521:  0.0764  0.0100 -0.0423  0.0107  0.0179 ...
}
\source{
\code{http://finance.yahoo.com/q/hp?s=GOOG}
}
\examples{
data(google)
## maybe str(google) ; plot(google) ...
}
\keyword{datasets}
