% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularization.R
\name{GroupBy}
\alias{GroupBy}
\title{GroupBy}
\usage{
GroupBy(arr, aggregation.function, n.columns.key = 1, n.columns.value = 1)
}
\arguments{
\item{arr}{TSA Array with the time series.}

\item{aggregation.function}{Function to be used in the aggregation. It receives an integer which indicates the
function to be applied:
{
  0 : mean,
  1 : median
  2 : min,
  3 : max,
  4 : stdev,
  5 : var,
  default : mean
}}

\item{n.columns.key}{Number of columns conforming the key.}

\item{n.columns.value}{Number of columns conforming the value (they are expected to be consecutive to the column
keys).}
}
\value{
TSA Array with the values of the group keys aggregated using the aggregation_function.
}
\description{
Group by operation in the input array using n.columns.key columns as group keys and n.columns.value columns as
values. The data is expected to be sorted. The aggregation function determines the operation to aggregate the values.
}
