% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{HasDuplicateMax}
\alias{HasDuplicateMax}
\title{HasDuplicateMax}
\usage{
HasDuplicateMax(tss)
}
\arguments{
\item{tss}{List of arrays of type double containing the time series.}
}
\value{
Array containing True if the maximum value of the time series is duplicated
and false otherwise.
}
\description{
Calculates if the maximum within input time series is duplicated.
}
