% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{HasDuplicateMin}
\alias{HasDuplicateMin}
\title{HasDuplicateMin}
\usage{
HasDuplicateMin(tss)
}
\arguments{
\item{tss}{List of arrays of type double containing the time series.}
}
\value{
List with an array containing True if the minimum of the time series is duplicated
and False otherwise.
}
\description{
Calculates if the minimum of the input time series is duplicated.
}
