\name{ima22.s}
\alias{ima22.s}
\docType{data}
\title{Simulated IMA(2,2) series / time series}
\description{A simulated IMA(2,2) series with theta1=1 and
theta2=-0.6}
\usage{data(ima22.s)}
\format{
  The format is:
 Time-Series [1:62] from 1 to 62:  0.00000  0.00000 -0.00569  2.12404  2.15337 ...
}
\examples{
data(ima22.s)
## maybe str(ima22.s) ; plot(ima22.s) ...
}
\keyword{datasets}
