% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{IndexMaxQuantile}
\alias{IndexMaxQuantile}
\title{IndexMaxQuantile}
\usage{
IndexMaxQuantile(tss, q)
}
\arguments{
\item{tss}{List of arrays of type double containing the time series.}

\item{q}{The quantile.}
}
\value{
The index of the max quantile q.
}
\description{
Calculates the index of the max quantile.
}
