% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{KurtosisStatistics}
\alias{KurtosisStatistics}
\title{KurtosisStatistics}
\usage{
KurtosisStatistics(tss)
}
\arguments{
\item{tss}{Expects an input array whose dimension zero is the length of the time series (all the same) and
dimension one indicates the number of time series.}
}
\value{
The kurtosis of tss.
}
\description{
Returns the kurtosis of tss (calculated with the adjusted Fisher-Pearson standardized moment coefficient G2).
}
