\name{larain}
\alias{larain}
\docType{data}
\title{Annual rainfall in Los Angeles / time series}
\description{
Annual precipitation (in inches) in Los Angeles, 1878-1992.
}
\usage{data(larain)}
\format{
  The format is:
 Time-Series [1:115] from 1778 to 1892: 20.86 17.41 18.65  5.53 10.74 ...
}
\source{
Personal communication from Professor Donald Bentley,
Pomona College, Claremont, California. For more data see
\code{http://www.wrh.noaa.gov/lox/climate/cvc.php}
}
\examples{
data(larain)
## maybe str(larain) ; plot(larain) ...
}
\keyword{datasets}
