% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{LargeStandardDeviation}
\alias{LargeStandardDeviation}
\title{LargeStandardDeviation}
\usage{
LargeStandardDeviation(tss, r)
}
\arguments{
\item{tss}{List of arrays of type double containing the time series.}

\item{r}{The threshold.}
}
\value{
Array containing True for those time series in tss that have a large standard deviation.
}
\description{
Checks if the time series within tss have a large standard deviation.
}
