% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{LastLocationOfMinimum}
\alias{LastLocationOfMinimum}
\title{LastLocationOfMinimum}
\usage{
LastLocationOfMinimum(arr)
}
\arguments{
\item{arr}{TSA array with the time series.}
}
\value{
The last relative location of the minimum value of each series.
}
\description{
Calculates the last location of the minimum value of each time series. The position
is calculated relatively to the length of the series.
}
