% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{Length}
\alias{Length}
\title{Length}
\usage{
Length(tss)
}
\arguments{
\item{tss}{List of arrays of type double containing the time series.}
}
\value{
The length of tss.
}
\description{
Returns the length of the input time series.
}
