% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{Linear}
\alias{Linear}
\title{Linear}
\usage{
Linear(xss, yss)
}
\arguments{
\item{xss}{TSA array with the time series.}

\item{yss}{TSA array with the time series.}
}
\value{
List with: {
        slope Slope of the regression line.
        intercept Intercept of the regression line.
        rvalue Correlation coefficient.
        pvalue Two-sided p-value for a hypothesis test whose null hypothesis is that the slope is zero, using Wald
                  Test with t-distribution of the test statistic.
        stderrest Standard error of the estimated gradient.
}
}
\description{
Calculates a linear least-squares regression for two sets of measurements. Both arrays should have the same
length.
}
