% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{LjungBox}
\alias{LjungBox}
\title{LjungBox}
\usage{
LjungBox(tss, lags)
}
\arguments{
\item{tss}{Expects an input array whose dimension zero is the length of the time series (all the same) and dimension
one indicates the number of time series.}

\item{lags}{Number of lags being tested.}
}
\value{
Array containing the Ljung-Box statistic test.
}
\description{
The Ljung–Box test checks that data whithin the time series are independently distributed (i.e. the correlations in
the population from which the sample is taken are 0, so that any observed correlations in the data result from
randomness of the sampling process). Data are no independently distributed, if they exhibit serial correlation.
}
