% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{LocalMaximals}
\alias{LocalMaximals}
\title{Calculates all Local Maximals fot the time series in array.}
\usage{
LocalMaximals(arr)
}
\arguments{
\item{arr:}{TSA array with the time series.}
}
\value{
: TSA array with the calculated local maximals for each time series in arr.
}
\description{
Calculates all Local Maximals fot the time series in array.
}
