% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{Maximum}
\alias{Maximum}
\title{Maximum}
\usage{
Maximum(tss)
}
\arguments{
\item{tss}{List of arrays of type double containing the time series.}
}
\value{
The maximum value of each time series within tss.
}
\description{
Calculates the maximum value for each time series within tss.
}
