% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{MaxLangevinFixedPoint}
\alias{MaxLangevinFixedPoint}
\title{MaxLangevinFixedPoint}
\usage{
MaxLangevinFixedPoint(tss, m, r)
}
\arguments{
\item{tss}{List of arrays of type double containing the time series.}

\item{m:}{Order of polynom to fit for estimating fixed points of dynamics.}

\item{r:}{Number of quantiles to use for averaging.}
}
\value{
Largest fixed point of deterministic dynamics.
}
\description{
Largest fixed point of dynamics \eqn{\max_x {h(x)=0}} estimated from polynomial
\eqn{h(x)}, which has been fitted to the deterministic dynamics of Langevin model
\deqn{
  \dot(x)(t) = h(x(t)) + R \mathcal(N)(0,1)
  }
as described by
Friedrich et al. (2000): Physics Letters A 271, p. 217-222 *Extracting model equations from experimental data.
}
