% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{MaxMinNormInPlace}
\alias{MaxMinNormInPlace}
\title{MaxMinNormInPlace}
\usage{
MaxMinNormInPlace(tss, high = 1, low = 1, epsilon = 1e-08)
}
\arguments{
\item{tss:}{TSA array with the time series.}

\item{high:}{Maximum final value (Defaults to 1.0).}

\item{low:}{Minimum final value (Defaults to 0.0).}

\item{epsilon:}{Safeguard for constant (or near constant) time series as the operation implies a unit scale
operation between min and max values in the tss.}
}
\description{
Same as max_min_norm, but it performs the operation in place, without allocating further memory.
}
