% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{Mean}
\alias{Mean}
\title{Mean}
\usage{
Mean(tss)
}
\arguments{
\item{tss}{List of arrays of type double containing the time series.}
}
\value{
The mean value of each time series within tss.
}
\description{
Calculates the mean value for each time series within tss.
}
