% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{MeanAbsoluteChange}
\alias{MeanAbsoluteChange}
\title{MeanAbsolutgeChange}
\usage{
MeanAbsoluteChange(tss)
}
\arguments{
\item{tss}{List of arrays of type double containing the time series.}
}
\value{
The mean over the absolute differences between subsequent time series values.
}
\description{
Calculates the mean over the absolute differences between subsequent time series values in tss.
}
