% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{MeanNormInPlace}
\alias{MeanNormInPlace}
\title{MeanNormInPlace}
\usage{
MeanNormInPlace(tss)
}
\arguments{
\item{tss}{TSA array with the time series.}
}
\description{
Normalizes the given time series according to its maximum-minimum value and its mean. It follows the following
formulae: 
\deqn{ \acute{x} = \frac{x - mean(x)}{max(x) - min(x)}.}
}
