% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{MeanSecondDerivativeCentral}
\alias{MeanSecondDerivativeCentral}
\title{MeanSecondDerivativeCentral}
\usage{
MeanSecondDerivativeCentral(tss)
}
\arguments{
\item{tss}{List of arrays of type double containing the time series.}
}
\value{
The mean value of a central approximation of
the second derivative for each time series.
}
\description{
Calculates mean value of a central approximation of the second derivative for each time series in tss.
}
