% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{Median}
\alias{Median}
\title{Median}
\usage{
Median(tss)
}
\arguments{
\item{tss}{List of arrays of type double containing the time series.}
}
\value{
The median value of each time series within tss.
}
\description{
Calculates the median value for each time series within tss.
}
