% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{Minimum}
\alias{Minimum}
\title{Minimum}
\usage{
Minimum(tss)
}
\arguments{
\item{tss}{List of arrays of type double containing the time series.}
}
\value{
The minimum value of each time series within tss.
}
\description{
Calculates the minimum value for each time series within tss.
}
