% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{NumberCrossingM}
\alias{NumberCrossingM}
\title{NumberCrossingM}
\usage{
NumberCrossingM(arr, m)
}
\arguments{
\item{arr}{TSA array with the time series.}

\item{m}{The m value.}
}
\value{
The number of m-crossings of each time series within arr.
}
\description{
Calculates the number of m-crossings. A m-crossing is defined as two sequential values where the first
value is lower than m and the next is greater, or viceversa. If you set m to zero, you will get the number of
zero crossings.
}
