% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{NumberCwtPeaks}
\alias{NumberCwtPeaks}
\title{NumberCwtPeaks}
\usage{
NumberCwtPeaks(arr, max.w)
}
\arguments{
\item{arr}{TSA array with the time series.}

\item{max.w}{The maximum width to consider.}
}
\value{
: TSA array with the number of peaks for each time series.
}
\description{
This feature calculator searches for different peaks. To do so, the time series is smoothed by a ricker
wavelet and for widths ranging from 1 to :math:'max_w`. This feature calculator returns the number of peaks that
occur at enough width scales and with sufficiently high Signal-to-Noise-Ratio (SNR).
}
