% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{PartialAutocorrelation}
\alias{PartialAutocorrelation}
\title{PartialAutocorrelation}
\usage{
PartialAutocorrelation(arr, lags)
}
\arguments{
\item{arr:}{TSA array with the time series.}

\item{lags:}{TSA array with the lags to be calculated.}
}
\value{
: TSA array with the partial autocorrelation for each time series for the given lag.
}
\description{
Calculates the value of the partial autocorrelation function at the given lag. The lag \eqn{k}  partial
autocorrelation of a time series \eqn{\lbrace x_t, t = 1 \ldots T \rbrace} equals the partial correlation of
\eqn{x_t} and \eqn{x_{t-k}}, adjusted for the intermediate variables \eqn{\lbrace x_{t-1}, \ldots, x_{t-k+1}
\rbrace} ([1]). Following [2], it can be defined as:
}
\details{
\deqn{
        \alpha_k = \frac{ Cov(x_t, x_{t-k} | x_{t-1}, \ldots, x_{t-k+1})}
        {\sqrt{ Var(x_t | x_{t-1}, \ldots, x_{t-k+1}) Var(x_{t-k} | x_{t-1}, \ldots, x_{t-k+1} )}}
}
with (a) \eqn{x_t = f(x_{t-1}, \ldots, x_{t-k+1})} and (b) \eqn{x_{t-k} = f(x_{t-1}, \ldots, x_{t-k+1})}
being AR(k-1) models that can be fitted by OLS. Be aware that in (a), the regression is done on past values to
predict \eqn{x_t} whereas in (b), future values are used to calculate the past value \eqn{x_{t-k}}.
It is said in [1] that "for an AR(p), the partial autocorrelations \eqn{\alpha_k} will be nonzero for \eqn{k<=p}
and zero for \eqn{k>p}."
With this property, it is used to determine the lag of an AR-Process.

[1] Box, G. E., Jenkins, G. M., Reinsel, G. C., & Ljung, G. M. (2015).
Time series analysis: forecasting and control. John Wiley & Sons.
[2] https://onlinecourses.science.psu.edu/stat510/node/62
}
