% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{PercentageOfReoccurringDatapointsToAllDatapoints}
\alias{PercentageOfReoccurringDatapointsToAllDatapoints}
\title{PercentageOfReoccurringDatapointsToAllDatapoints}
\usage{
PercentageOfReoccurringDatapointsToAllDatapoints(arr, is.sorted)
}
\arguments{
\item{arr}{TSA array with the time series.}

\item{is.sorted}{Indicates if the input time series is sorted or not. Defaults to false..}
}
\value{
TSA array with the percentage of unique data points, that are present in the time series more than once.
}
\description{
Calculates the percentage of unique data points, that are present in the time series more than once.
\deqn{
  \frac{\textit{number of data points occurring more than once}}{\textit{number of all data points})}
}
}
