% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{PercentageOfReoccurringValuesToAllValues}
\alias{PercentageOfReoccurringValuesToAllValues}
\title{PercentageOfReoccurringValuesToAllValues}
\usage{
PercentageOfReoccurringValuesToAllValues(arr, is.sorted = FALSE)
}
\arguments{
\item{arr}{TSA array with the time series.}

\item{is.sorted}{Indicates if the input time series is sorted or not. Defaults to false.}
}
\value{
: TSA array with the percentage of unique values, that are present in the time series more than once.
}
\description{
Calculates the percentage of unique values, that are present in the time series more than once.
}
\details{
\deqn{
  \frac{\textit{number of values occurring more than once}}{\textit{number of all values})}
}

This means the percentage is normalized to the number of unique values, in contrast to the
PercentageOfReoccurringDatapointsToAllDatapoints.
}
