% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality.R
\name{PLASlidingWindow}
\alias{PLASlidingWindow}
\title{PLASlidingWindow}
\usage{
PLASlidingWindow(ts, max.error)
}
\arguments{
\item{ts}{Expects a tsa array containing the set of points to be reduced. The first component of the points in
the first column and the second component of the points in the second column.}

\item{max.error}{The maximum approximation error allowed.}
}
\value{
The reduced number of points.
}
\description{
Applies the Piecewise Linear Approximation (PLA SlidingWindow) to the time series.
}
\details{
[1] Zhu Y, Wu D, Li Sh (2007). A Piecewise Linear Representation Method of Time Series Based on Feature Points.
Knowledge-Based Intelligent Information and Engineering Systems 4693:1066-1072.
}
