\name{plot1.acf}
\alias{plot1.acf}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Plot1}
\description{
A modification of the plot.acf function in the stats pacakage, mainly
as a workhorse function for the acf function in the TSA pacakage.}
\author{Kung-Sik Chan}
\keyword{misc}
