% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial.R
\name{Polyfit}
\alias{Polyfit}
\title{Polyfit}
\usage{
Polyfit(x, y, deg)
}
\arguments{
\item{x:}{TSA array with the x-coordinates of the M sample points \eqn{(x[i], y[i])}.}

\item{y:}{TSA array with the y-coordinates of the sample points.}

\item{deg:}{Degree of the fitting polynomial}
}
\value{
: TSA array with the polynomial coefficients, highest power first.
}
\description{
Least squares polynomial fit. Fit a polynomial \eqn{p(x) = p[0] * x^{deg} + ... + p[deg]} of degree
\eqn{deg} to points \eqn{(x, y)}. Returns a vector of coefficients \eqn{p} that minimises the squared error.
}
