\name{prey.eq}
\alias{prey.eq}
\docType{data}
\title{Prey series / time series}
\description{
The stationary part of the Didinium series in the veilleux data frame.
}
\usage{data(prey.eq)}
\format{
  The format is:
 Time-Series [1:57] from 7 to 35:  26.9  53.2  65.6  81.2 143.9 ...
}
\seealso{ \code{\link{veilleux}}}
\examples{
data(prey.eq)
## maybe str(prey.eq) ; plot(prey.eq) ...
}
\keyword{datasets}
