% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{Quantile}
\alias{Quantile}
\title{Quantile}
\usage{
Quantile(arr, arr.q, precision = 1e+08)
}
\arguments{
\item{arr}{TSA array with the time series.}

\item{precision}{Number of decimals expected.}

\item{q}{Percentile(s) at which to extract score(s). One or many.}
}
\value{
The number of m-crossings of each time series within arr.
}
\description{
Returns values at the given quantile.
}
