% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{QuantilesCutStatistics}
\alias{QuantilesCutStatistics}
\title{QuantilesCutStatistics}
\usage{
QuantilesCutStatistics(tss, q, precision = 1e-08)
}
\arguments{
\item{tss}{Expects an input array whose dimension zero is the length of the time series (all the same) and
dimension one indicates the number of time series.}

\item{q}{Number of quantiles to extract. From 0 to 1, step 1/quantiles.}

\item{precision}{Number of decimals expected.}
}
\value{
Matrix with the categories, one category per row, the start of the category in the first column and
the end in the second category.
}
\description{
Discretizes the time series into equal-sized buckets based on sample quantiles.
}
