% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality.R
\name{RamerDouglasPeucker}
\alias{RamerDouglasPeucker}
\title{RamerDouglasPeucker}
\usage{
RamerDouglasPeucker(arr, epsilon)
}
\arguments{
\item{arr}{TSA Array with with the x-coordinates and y-coordinates of the input points (x in column 0 and y in column 1).}

\item{epsilon}{It acts as the threshold value to decide which points should be considered meaningful or not.}
}
\value{
Array with the x-coordinates and y-coordinates of the selected points (x in column 0 and y in
column 1).
}
\description{
The Ramer–Douglas–Peucker algorithm (RDP) is an algorithm for reducing the number of points in a curve that is
approximated by a series of points. It reduces a set of points depending on the perpendicular distance of the points
and epsilon, the greater epsilon, more points are deleted.
}
\details{
[1] Urs Ramer, "An iterative procedure for the polygonal approximation of plane curves", Computer Graphics and Image
Processing, 1(3), 244–256 (1972) doi:10.1016/S0146-664X(72)80017-0.

[2] David Douglas & Thomas Peucker, "Algorithms for the reduction of the number of points required to represent a
digitized line or its caricature", The Canadian Cartographer 10(2), 112–122 (1973) doi:10.3138/FM57-6770-U75U-7727
}
