% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{RangeCount}
\alias{RangeCount}
\title{RangeCount}
\usage{
RangeCount(arr, min, max)
}
\arguments{
\item{arr:}{TSA array with the time series.}

\item{min:}{Value that sets the lower limit.}

\item{max:}{Value that sets the upper limit.}
}
\value{
: TSA array with the values at the given range.
}
\description{
Counts observed values within the interval [min, max).
}
