% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{RatioBeyondRSigma}
\alias{RatioBeyondRSigma}
\title{RatioBeyondRSigma}
\usage{
RatioBeyondRSigma(arr, r)
}
\arguments{
\item{arr}{TSA array with the time series.}

\item{r}{Number of times that the values should be away from.}
}
\value{
TSA array with the ratio of values that are more than \eqn{r*std(x)} (so \eqn{r} sigma) away from
the mean of \eqn{x}.
}
\description{
Calculates the ratio of values that are more than  \eqn{r*std(x)} (so \eqn{r} sigma away from the mean of
\eqn{x}.
}
