% --- Source file: man/retail.Rd ---
\name{retail}
\alias{retail}
\docType{data}
\title{U.K. retail sales / time series}
\description{
Monthly total UK (United Kingdom) retail sales (non-food stores in
billions of pounds), 01/1983 - 12/1987.
}
\usage{data(retail)}
\format{
  The format is:
 Time-Series [1:60] from 1983 to 1988: 81.3 78.9 93.8 94 97.8 1.6 99.6 1.2 98 1.7 ...
}
\source{
www.statistics.gov.uk/statbase/TSDdownload1.asp
}
\examples{
data(retail)
## maybe str(retail) ; plot(retail) ...
}
\keyword{datasets}
