\name{robot}
\alias{robot}
\docType{data}
\title{The distance of a robot from a desired position / time series}
\description{
Final position in the x direction of an industrial robot put through a
series of planned exercises many times.}
\usage{data(robot)}
\format{
  The format is:
 Time-Series [1:324] from 1 to 324: 0.0011 0.0011 0.0024 0 -0.0018 0.0055 0.0055 -0.0015 0.0047 -1e-04 ...
}
\source{
Personal communication
from William F. Fulkerson, Deere & Co. Technical Center,
Moline, Illinois.
}
\examples{
data(robot)
## maybe str(robot) ; plot(robot) ...
}
\keyword{datasets}
