% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial.R
\name{Roots}
\alias{Roots}
\title{Roots}
\usage{
Roots(p)
}
\arguments{
\item{p:}{TSA array with the polynomial coefficients.}
}
\value{
: TSA array with the roots of the polynomial.
}
\description{
Calculates the roots of a polynomial with coefficients given in \eqn{p}. The values in the rank-1 array
\eqn{p} are coefficients of a polynomial. If the length of \eqn{p} is \eqn{n+1} then the polynomial is
described by:
\deqn{
p[0] * x^n + p[1] * x^{n-1} + ... + p[n-1] * x + p[n]
}
}
