% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality.R
\name{Sax}
\alias{Sax}
\title{Sax}
\usage{
Sax(arr, alphabetSize)
}
\arguments{
\item{arr}{TSA Array with the input time series.}

\item{alphabetSize}{Number of element within the alphabet.}
}
\value{
Array of symbols.
}
\description{
Symbolic Aggregate approXimation (SAX). It transforms a numeric time series into a time series of symbols with
the same size. The algorithm was proposed by Lin et al.) and extends the PAA-based approach inheriting the original
algorithm simplicity and low computational complexity while providing satisfactory sensitivity and selectivity in
range query processing. Moreover, the use of a symbolic representation opened a door to the existing wealth of
data-structures and string-manipulation algorithms in computer science such as hashing, regular expression, pattern
matching, suffix trees, and grammatical inference.
}
\details{
[1] Lin, J., Keogh, E., Lonardi, S. & Chiu, B. (2003) A Symbolic Representation of Time Series, with Implications for
Streaming Algorithms. In proceedings of the 8th ACM SIGMOD Workshop on Research Issues in Data Mining and Knowledge
Discovery. San Diego, CA. June 13.
}
