\name{SP}
\alias{SP}
\docType{data}
\title{Quarterly Standard & Poor's Composite Index of stock price values /
time series}
\description{
Quarterly S&P Composite Index, 1936Q1 - 1977Q4.
}
\usage{data(SP)}
\format{
  The format is:
 Time-Series [1:168] from 1936 to 1978: 149 148 160 172 179 ...
}
\source{
Frees,
E. W., Data Analysis Using Regression Models, Prentice Hall, 1996.
}
\examples{
data(SP)
## maybe str(SP) ; plot(SP) ...
}
\keyword{datasets}
