% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{SquaredEuclidean}
\alias{SquaredEuclidean}
\title{SquaredEuclidean}
\usage{
SquaredEuclidean(tss)
}
\arguments{
\item{tss}{Expects an input array whose dimension zero is the length of the time series (all the same) and
dimension one indicates the number of time series.}
}
\value{
TSA array with an upper triangular matrix where each position corresponds to the distance between two
time series. Diagonal elements will be zero. For example: Position row 0 column 1 records the distance
between time series 0 and time series 1.
}
\description{
Calculates the non squared version of the euclidean distance.
}
