% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{stamp}
\alias{stamp}
\title{Stamp algorithm
STAMP algorithm to calculate the matrix profile between 'ta' and 'tb' using a subsequence length
of 'm'.}
\usage{
stamp(first_time_series, second_time_series, subsequence_length)
}
\value{
A matrix profile
}
\description{
Stamp algorithm
STAMP algorithm to calculate the matrix profile between 'ta' and 'tb' using a subsequence length
of 'm'.
}
