% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_self_join.R
\name{stamp_self_join}
\alias{stamp_self_join}
\title{Stamp self join algorithm
STAMP algorithm to calculate the matrix profile between 't' and itself using a subsequence length
of 'm'. This method filters the trivial matches.}
\usage{
stamp_self_join(first_time_series, subsequence_length)
}
\value{
A matrix profile
}
\description{
Stamp self join algorithm
STAMP algorithm to calculate the matrix profile between 't' and itself using a subsequence length
of 'm'. This method filters the trivial matches.
}
