% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{Stomp}
\alias{Stomp}
\title{Stomp}
\usage{
Stomp(first.time.series, second.time.series, subsequence.length)
}
\arguments{
\item{second.time.series}{List of arrays of type double containing the time series.}

\item{firs.time.series}{List of arrays of type double containing the time series.}

\item{subsequence.lenth}{Length of the subsequence.}
}
\value{
A matrix profile
}
\description{
STOMP algorithm to calculate the matrix profile between 'ta' and 'tb' using a subsequence length
of 'm'.
}
