% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stomp_self_join.R
\name{stomp_self_join}
\alias{stomp_self_join}
\title{Stomp self join algorithm
STOMP algorithm to calculate the matrix profile between 't' and itself using a subsequence length
of 'm'. This method filters the trivial matches.}
\usage{
stomp_self_join(first_time_series, subsequence_length)
}
\value{
A matrix profile
}
\description{
Stomp self join algorithm
STOMP algorithm to calculate the matrix profile between 't' and itself using a subsequence length
of 'm'. This method filters the trivial matches.
}
