% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{StompSelfJoin}
\alias{StompSelfJoin}
\title{StompSelfJoin}
\usage{
StompSelfJoin(first.time.series, subsequence.length)
}
\arguments{
\item{first.time.series}{List of arrays of type double containing the time series.}

\item{subsequence.length}{Lenght of the subsequence}
}
\value{
A matrix profile
}
\description{
STOMP algorithm to calculate the matrix profile between 't' and itself using a subsequence length
of 'm'. This method filters the trivial matches.
}
