% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{SumOfReoccurringDatapoints}
\alias{SumOfReoccurringDatapoints}
\title{SumOfReoccuringDatapoints}
\usage{
SumOfReoccurringDatapoints(arr, is.sorted = FALSE)
}
\arguments{
\item{arr}{TSA array with the time series.}
}
\value{
Returns the sum of all data points, that are present in the time series more than once.
}
\description{
Calculates the sum of all data points, that are present in the time series more than once.
}
