% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{SumValues}
\alias{SumValues}
\title{SumValues}
\usage{
SumValues(arr)
}
\arguments{
\item{arr:}{TSA array with the time series.}
}
\value{
: TSA array with the sum of values in each time series.
}
\description{
Calculates the sum over the time series arr.
}
