% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{SymmetryLooking}
\alias{SymmetryLooking}
\title{SymmetryLooking}
\usage{
SymmetryLooking(arr, r)
}
\arguments{
\item{arr}{TSA array with the time series.}

\item{r}{The percentage of the range to compare with.}
}
\value{
An array denoting if the input time series look symmetric.
}
\description{
Calculates if the distribution of arr *looks symmetric*. This is the case if
\deqn{
 | mean(arr)-median(arr)| < r * (max(arr)-min(arr))
}
}
