\name{tbone}
\alias{tbone}
\docType{data}
\title{
A digitized sound file of a B flat 
played on a tenor trombone}
\description{
A digitized sound file of about 0.4 seconds of a B flat just below
middle C played on a tenor trombone by Chuck Kreeb, a member of
Tempered Brass and a friend of one of the authors.
}
\usage{data(tbone)}
\format{
  The format is:
 Time-Series [1:17689] from 1 to 17689: 0.0769 0.0862 0.0961 0.1050 0.1129 ...
}
\examples{
data(tbone)
## maybe str(tbone) ; plot(tbone) ...
}
\keyword{datasets}
