% --- Source file: man/tempdub.Rd ---
\name{tempdub}
\alias{tempdub}
\docType{data}
\title{Monthly average temperature in Dubuque/time series}
\description{Monthly average temperature (in degrees Fahrenheit) 
recorded in Dubuque 1/1964 - 12/1975.}
\usage{data(tempdub)}
\format{
  The format is:
 Time-Series [1:144] from 1964 to 1976: 24.7 25.7 30.6 47.5 62.9 68.5 73.7 67.9 61.1 48.5 ...
}
\source{
\code{http://mesonet.agron.iastate.edu/climodat/index.phtml?station=ia2364&report=16}}
\examples{
data(tempdub)
## maybe str(tempdub) ; plot(tempdub) ...
}
\keyword{datasets}
