% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{TimeReversalAsymmetryStatistic}
\alias{TimeReversalAsymmetryStatistic}
\title{TimeReversalAsymmetryStatistic}
\usage{
TimeReversalAsymmetryStatistic(arr, lag)
}
\arguments{
\item{arr}{TSA array with the time series.}

\item{lag}{The lag to be computed.}
}
\value{
TSA array containing the count of the given value in each time series.
}
\description{
This function calculates the value of:
}
\details{
\deqn{
        \frac{1}{n-2lag} \sum_{i=0}^{n-2lag} x_{i + 2 \cdot lag}^2 \cdot x_{i + lag} - x_{i + lag} \cdot  x_{i}^2
}
which is

\deqn{
       \mathbb{E}[L^2(X)^2 \cdot L(X) - L(X) \cdot X^2]
}
where \eqn{\mathbb{E}} is the mean and \eqn{L} is the lag operator. It was proposed in [1] as a promising
feature to extract from time series.
}
