\name{tuba}
\alias{tuba}
\docType{data}
\title{
A digitized sound file of a B flat 
played on a BB flat tuba}
\description{
A digitized sound file of about 0.4 seconds of a B flat an octave and
one whole step below middle C played on a BB flat tuba by Linda
Fisher, a member of Tempered Brass and a friend one of the authors.
}
\usage{data(tuba)}
\format{
  The format is:
 Time-Series [1:4402] from 1 to 4402: 0.217 0.209 0.200 0.195 0.196 ...
}
\examples{
data(tuba)
## maybe str(tuba) ; plot(tuba) ...
}
\keyword{datasets}
