\name{units}
\alias{units}
\docType{data}
\title{
Annual sales of certain large equipment
}
\description{
Annual sales of certain large equipment, 1983 - 2005.
}
\usage{data(units)}
\format{
  The format is:
 ts [1:24, 1]  71.7  78.6 111.1 125.6 133.0 ...
 - attr(*, "tsp")= num [1:3] 1982 2005    1
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr "Units"
}
\source{
Proprietary
sales data from a large international company
}
\examples{
data(units)
## maybe str(units) ; plot(units) ...
}
\keyword{datasets}
