% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{VarianceLargerThanStandardDeviation}
\alias{VarianceLargerThanStandardDeviation}
\title{Variance Larger Than Standard Deviation}
\usage{
VarianceLargerThanStandardDeviation(arr)
}
\arguments{
\item{arr}{TSA array with the time series.}
}
\value{
TSA array denoting if the variance of array is greater than the standard deviation.
}
\description{
Calculates if the variance of array is greater than the standard deviation. In other words, if the variance of
array is larger than 1.
}
