% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality.R
\name{Visvalingam}
\alias{Visvalingam}
\title{Visvalingam}
\usage{
Visvalingam(arr, numPoints)
}
\arguments{
\item{arr}{TSA Array with with the x-coordinates and y-coordinates of the input points (x in column 0 and y in column 1).}

\item{numPoints}{Sets the number of points returned after the execution of the method.}
}
\value{
Array with the x-coordinates and y-coordinates of the selected points (x in column 0 and y in
column 1).
}
\description{
Reduces a set of points by applying the Visvalingam method (minimun triangle area) until the number
of points is reduced to numPoints.
'
[1] M. Visvalingam and J. D. Whyatt, Line generalisation by repeated elimination of points,
The Cartographic Journal, 1993.
}
