% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\docType{methods}
\name{xor.tsa,Array,Array-method}
\alias{xor.tsa,Array,Array-method}
\title{xor.tsa}
\usage{
\S4method{xor.tsa}{Array,Array}(e1, e2)
}
\description{
Returns a exclusive-or between each pair of elements from two TSA arrays (element-wise).
}
