% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{Znorm}
\alias{Znorm}
\title{Znorm}
\usage{
Znorm(arr, epsilon = 1e-08)
}
\arguments{
\item{arr}{TSA Array with the time series.}

\item{epsilon}{Minimum standard deviation to consider. It acts a a gatekeeper for
those time series that may be constant or near constant.}
}
\value{
Array with the same dimensions as arr where the time series have been
adjusted for zero mean and one as standard deviation.
}
\description{
Calculates a new set of time series with zero mean and standard deviation one.
}
