% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{ZnormInPlace}
\alias{ZnormInPlace}
\title{ZnormInPLace}
\usage{
ZnormInPlace(arr, epsilon = 1e-08)
}
\arguments{
\item{arr:}{TSA array with the time series.}

\item{epsilon:}{epsilon Minimum standard deviation to consider.  It acts as a gatekeeper for
those time series that may be constant or near constant.}
}
\description{
Adjusts the time series in the given input and performs z-norm
inplace (without allocating further memory).
}
