% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{AbsEnergy}
\alias{AbsEnergy}
\title{AbsEnergy}
\usage{
AbsEnergy(time.series)
}
\arguments{
\item{time.series}{List of arrays of type double containing the time series.}
}
\value{
List with the Absolute Energy.
}
\description{
Calculates de sum over the square values of the time series.
}
