% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{AbsoluteSumOfChanges}
\alias{AbsoluteSumOfChanges}
\title{AbsoluteSumOfChanges}
\usage{
AbsoluteSumOfChanges(arr)
}
\arguments{
\item{arr}{TSA array with the time series.}
}
\value{
TSA Array with the AbsoluteSumOfChanges
}
\description{
Calculates the sum over the absolute value of consecutive
changes in the time series.
}
