% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{AggregatedAutocorrelation}
\alias{AggregatedAutocorrelation}
\title{AggregatedAutocorrelation}
\usage{
AggregatedAutocorrelation(arr, aggregation.function)
}
\arguments{
\item{arr}{TSA array with the time series.}

\item{aggregation.function}{Function to be used in the aggregation. It receives an integer which indicates the
function to be applied:
{
  0 : mean,
  1 : median
  2 : min,
  3 : max,
  4 : stdev,
  5 : var,
  default : mean
}}
}
\value{
TSA Array whose values contains the aggregated correlation for each time series.
}
\description{
Calculates a linear least-squares regression for values of the time series that were aggregated
over chunks versus the sequence from 0 up to the number of chunks minus one.
}
