% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{Array}
\alias{Array}
\title{Array}
\usage{
Array(a)
}
\arguments{
\item{a}{data.frame which columns have same types and same dimensions.}
}
\value{
A TSA array created from the data stored in the data.frame.
}
\description{
Creates a tsa Array from a data.frame.
Every columns of the data.frame represents an array and every one should have the same type and same dimensions
}
