\name{beersales}
\alias{beersales}
\docType{data}
\title{Monthly beer sales / time series}
\description{
Monthly beer sales in millions of barrels, 01/1975 - 12/1990.
}
\usage{data(beersales)}
\format{
  The format is:
 Time-Series [1:192] from 1975 to 1991: 11.12  9.84 11.57 13.01 13.42 ...
}
\source{
Frees, E. W., Data Analysis Using Regression Models,
Prentice Hall, 1996.
}
\examples{
data(beersales)
## maybe str(beersales) ; plot(beersales) ...
}
\keyword{datasets}
