\name{bluebirdlite}
\alias{bluebirdlite}
\docType{data}
\title{
Bluebird Lite potato chip data
}
\description{
Weekly unit sales (log-transformed) of Bluebird Lite potato chips (New Zealand) and
their price for 104 weeks. 
}
\usage{data(bluebirdlite)}
\format{
  A data frame with 104 observations on the following 2 variables.
  \describe{
    \item{\code{log.sales}}{a numeric vector}
    \item{\code{price}}{a numeric vector}
  }
}
\source{
www.stat.auckland.ac.nz/~balemi/Assn3.xls
}
\examples{
data(bluebirdlite)
## maybe str(bluebirdlite) ; plot(bluebirdlite) ...
}
\keyword{datasets}
