\name{boardings}
\alias{boardings}
\docType{data}
\title{
Monthly public transit boardings and gasoline price in Denver}

\description{
Monthly public transit boardings (mostly buses and light rail) and
gasoline price (both log-transformed), Denver,
Colorado region, 08/2000 - 03/2006.
}
\source{
Personal communication
from Lee Cryer, Project Manager, Regional Transportation
District, Denver, Colorado. Denver gasoline prices were obtained
from the Energy Information Administration, U.S. Department of
Energy, Washington, D.C.at www.eia.doe.gov

}
\examples{
data(boardings)
plot(boardings)
## maybe str(boardings) ; plot(boardings) ...
}
\keyword{datasets}
