\name{co2}
\alias{co2}
\docType{data}
\title{Levels of Carbon Dioxide at Alert, Canada / Time series}  
\description{Monthly CO2 level at Alert, Northwest Territories, Canada,
near the Artic Circle, 01/1994 - 12/2004.}
\usage{data(co2)}
\format{
  The format is:
 Time-Series [1:132] from 1994 to 2005: 363 364 365 364 364 ...
}
\source{\code{http://cdiac.ornl.gov/trends/co2/sio-alt.htm}}
\examples{
data(co2)
## maybe str(co2) ; plot(co2) ...
}
\keyword{datasets}
