\name{cref.bond}
\alias{cref.bond}
\docType{data}
\title{Daily CREF Bond Values}
\description{
Daily values of one unit of the CREF (College Retirement Equity
Fund) Bond fund, 08/26/04 - 08/15/06.}
\usage{data(CREF)}
\source{
www.tiaa-cref.org/performance/retirement/data/index.html
}
\examples{
data(CREF)
## maybe str(CREF) ; plot(CREF) ...
}
\keyword{datasets}
