\name{CREF}
\alias{CREF}
\docType{data}
\title{Daily CREF Values}
\description{
Daily values of one unit of the CREF (College Retirement Equity
Fund) Stock fund, 08/26/04 - 08/15/06.
}
\usage{data(CREF)}
\format{
  The format is:
 Time-Series [1:501] from 1 to 501: 170 170 169 170 171 ...
}
\source{
www.tiaa-cref.org/performance/retirement/data/index.html
}
\examples{
data(CREF)
## maybe str(CREF) ; plot(CREF) ...
}
\keyword{datasets}
