% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{CrossCovariance}
\alias{CrossCovariance}
\title{CrossCovariance}
\usage{
CrossCovariance(xss, yss, unbiased)
}
\arguments{
\item{xss}{List of arrays of type double containing the time series.}

\item{yss}{List of arrays of type double containing the time series.}

\item{unbiased}{Determines whether it divides by n - lag (if true) or n (if false).}
}
\value{
The cross-covariance value for the given time series.
}
\description{
Calculates the cross-covariance of the given time series
}
