% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{CwtCoefficients}
\alias{CwtCoefficients}
\title{CwtCoefficients
Calculates a Continuous wavelet transform for the Ricker wavelet, also known as
the "Mexican hat wavelet".}
\usage{
CwtCoefficients(tss, widths, coeff, w)
}
\arguments{
\item{tss}{List of arrays of type double containing the time series.}

\item{widths}{Widths. List of arrays of type double containing the time series.}

\item{coeff}{Coefficient of interest.}

\item{w}{Width of interest.}
}
\value{
Result of calculated coefficients.
}
\description{
CwtCoefficients
Calculates a Continuous wavelet transform for the Ricker wavelet, also known as
the "Mexican hat wavelet".
}
