\name{deere3}
\alias{deere3}
\docType{data}
\title{Deviations of an industrial process at Deere \& Co. -- Series 3} 
\description{
Fifty seven consecutive values for the deviation (in 0.0000025 inch units) from a specified target value.}
\usage{data(deere3)}
\format{
  The format is:
 Time-Series [1:57] from 1 to 57:  -500 -1250  -500 -3000 -2375 ...
}
\source{
Personal communication from William F.
Fulkerson, Deere \& Co. Technical Center, Moline, Illinois.
}
\examples{
data(deere3)
## maybe str(deere3) ; plot(deere3) ...
}
\keyword{datasets}
