% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(a)
}
\arguments{
\item{a}{TSA array of interest.}
}
\value{
The data stored by the TSA array in a vector of its corresponding R DATA type.
}
\description{
Gets the data to the host.
}
