\name{hare}
\alias{hare}
\title{Canadian hare data/ time series}
\description{Annual number of hare data.}
\usage{data(hare)}
\format{
  The format is:
 Time-Series [1:31] from 1905 to 1935: 50 20 20 22 27 50 55 78 70 59 ...
}
\details{
These are yearly hare abundances  for the main drainage of
the Hudson Bay, based 
on trapper questionnaires.
}
\source{
MacLulich, D. A.   (1937)   Fluctuations in the Number of the
Varying Hare (Lepus americanus)  (Univ. of Toronto Press, Toronto)
}
\references{
Stenseth, N. C., Falck, W., Bjornstad, O. N. and  Krebs. C. J. (1997)
Population regulation in snowshoe hare and Canadian lynx: Asymmetric
food web configurations between hare and lynx. Proc. Natl. Acad. Sci.,
94, 5147-5152.
}
\examples{
data(hare)
}
\keyword{datasets}
