\name{hours}
\alias{hours}
\docType{data}
\title{ Average hours worked in US manufacturing sector / time series}
\description{
 Average hours worked (times 10) in U.S. manufacturing sector,
from 07/1982 - 06/1987}
\usage{data(hours)}
\format{
  The format is:
 Time-Series [1:60] from 1983 to 1987: 389 390 389 390 393 397 392 388 396 398 ...
}
\source{
Cryer, J. D. Time Series Analysis,
Duxbury Press, 1986.
}
\examples{
data(hours)
## maybe str(hours) ; plot(hours) ...
}
\keyword{datasets}
