\name{JJ}
\alias{JJ}
\docType{data}
\title{Quarterly earnings per share for the Johnson \& Johnson Company}
\description{Quarterly earnings per share for 1960Q1 to 1980Q4 of the U.S. company,
Johnson \& Johnson, Inc.}
\usage{data(JJ)}
\format{
  The format is:
 Time-Series [1:84] from 1960 to 1981: 0.71 0.63 0.85 0.44 0.61 0.69 0.92 0.55 0.72 0.77 ...
}
\source{
\code{http://www.stat.pitt.edu/stoffer/tsa2/}
}
\examples{
data(JJ)
## maybe str(JJ) ; plot(JJ) ...
}
\keyword{datasets}
