% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{Kurtosis}
\alias{Kurtosis}
\title{Kurtosis}
\usage{
Kurtosis(tss)
}
\arguments{
\item{tss}{List of arrays of type double containing the time series.}
}
\value{
The kurtosis of each tss.
}
\description{
Returns the kurtosis of tss (calculated with the adjusted Fisher-Pearson
standardized moment coefficient G2).
}
