% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{Length}
\alias{Length}
\title{Length}
\usage{
Length(arr)
}
\arguments{
\item{arr}{TSA array with the time series.}
}
\value{
The length of arr.
}
\description{
Returns the length of the input time series.
}
