% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{LongestStrikeAboveMean}
\alias{LongestStrikeAboveMean}
\title{LongestStrikeAboveMean}
\usage{
LongestStrikeAboveMean(tss)
}
\arguments{
\item{tss}{List of arrays of type double containing the time series.}
}
\value{
The length of the longest consecutive subsequence in the input time series that is bigger than the mean.
}
\description{
Calculates the length of the longest consecutive subsequence in tss that is bigger than the mean of tss.
}
