% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{Maximum}
\alias{Maximum}
\title{Maximum}
\usage{
Maximum(arr)
}
\arguments{
\item{arr}{TSA Array of arrays of type double containing the time series.}
}
\value{
The maximum value of each time series within arr.
}
\description{
Calculates the maximum value for each time series within arr.
}
