\name{oil.price}
\alias{oil.price}
\docType{data}
\title{Monthly Oil Price / time series}
\description{
Monthly spot price for crude oil, Cushing, OK (in U.S. dollars per
barrel), 01/1986 - 01/2006.
}
\usage{data(oil.price)}
\format{
  The format is:
 Time-Series [1:241] from 1986 to 2006: 22.9 15.4 12.6 12.8 15.4 ...
}
\source{
tonto.eia.doe.gov/dnav/pet/hist/rwtcM.htm
}
\examples{
data(oil.price)
## maybe str(oil.price) ; plot(oil.price) ...
}
\keyword{datasets}
