\name{oilfilters}
\alias{oilfilters}
\docType{data}
\title{
Monthly sales to dealers of a specialty oil
filter/time series}
\description{
Monthly wholesale specialty oil filters sales, Deere & Co, 07/1983 -
06/1987.
}
\usage{data(oilfilters)}
\format{
  The format is:
 Time-Series [1:48] from 1984 to 1987: 2385 3302 3958 3302 2441 3107 5862 4536 4625 4492 ...
 - attr(*, "freq")= num 12
 - attr(*, "start")= num [1:2] 1987    7
}
\source{
Data courtesy of William F. Fulkerson, Deere & Company, Technical Center, Moline,
Illinois.}
\examples{
data(oilfilters)
## maybe str(oilfilters) ; plot(oilfilters) ...
}
\keyword{datasets}
