\name{prescrip}
\alias{prescrip}
\docType{data}
\title{ Cost per prescription / time series}
\description{
Monthly U.S. average prescription costs for the months 08/1986 -
03/1992.
}
\usage{data(prescrip)}
\format{
  The format is:
 Time-Series [1:68] from 1987 to 1992: 14.5 14.7 14.8 14.6 14.3 ...
}
\source{
Frees, E. W., Data Analysis Using Regression
Models, Prentice Hall, 1996.}
\examples{
data(prescrip)
## maybe str(prescrip) ; plot(prescrip) ...
}
\keyword{datasets}
