% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{RatioBeyondRSigma}
\alias{RatioBeyondRSigma}
\title{RatioBeyondRSigma}
\usage{
RatioBeyondRSigma(arr, r)
}
\arguments{
\item{arr}{TSA array with the time series.}

\item{q}{Percentile(s) at which to extract score(s). One or many.}

\item{precision}{Number of decimals expected.}
}
\value{
The number of m-crossings of each time series within arr.
}
\description{
Calculates the ratio of values that are more than  \eqn{r*std(x)} (so \eqn{r} sigma away from the mean of
\eqn{x}.
}
