% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{Stomp}
\alias{Stomp}
\title{Stomp}
\usage{
Stomp(first.time.series, second.time.series, subsequence.length)
}
\arguments{
\item{first.time.series}{TSA Array the time series.}

\item{second.time.series}{TSA Array time series.}

\item{subsequence.length}{Length of the subsequence.}
}
\value{
List of TSA Arrays with the matrix profile and the index profile
}
\description{
STOMP algorithm to calculate the matrix profile between 'first.time.series' and 'second.time.series' using a subsequence length
of 'subsequence.length'.
}
