% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stomp.R
\name{stomp}
\alias{stomp}
\title{Stomp algorithm
STOMP algorithm to calculate the matrix profile between 'ta' and 'tb' using a subsequence length
of 'm'.}
\usage{
stomp(first_time_series, second_time_series, subsequence_length)
}
\value{
A matrix profile
}
\description{
Stomp algorithm
STOMP algorithm to calculate the matrix profile between 'ta' and 'tb' using a subsequence length
of 'm'.
}
