\name{TSA-package}
\alias{TSA-package}
\alias{TSA}
\docType{package}
\title{Time Series Analysis}
\description{
Contains R functions and datasets detailed in the book
"Time Series Analysis with Applications in R (second edition)" by J.D. Cryer and K.S. Chan}
\details{
\tabular{ll}{
Package: \tab TSA\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2020-9-9\cr
License:\tab GPL version 2 or newer\cr
}
}
\author{Kung-Sik Chan, Brian Ripley}
\keyword{package}
