\name{TSA-package}
\alias{TSA-package}
\alias{TSA}
\docType{package}
\title{Time Series Analysis}
\description{
Contains R functions and datasets detailed in the book
"Time Series Analysis with Applications in R (second edition)" by J.D. Cryer and K.S. Chan}
\details{
\tabular{ll}{
Package: \tab TSA\cr
Type: \tab Package\cr
Version: \tab 0.97\cr
Date: \tab 2008-7-21\cr
License:\tab GPL version 2 or newer\cr
}
}
\author{Kung-Sik Chan
Maintainer: Kung-Sik Chan <kchan@stat.uiowa.edu>
}
\keyword{ package }
