\name{wages}
\alias{wages}
\docType{data}
\title{Average hourly wages in the apparel industry / time series}
\description{
Average hourly wages in the apparel industry, from 07/1981 - 06/1987.  
}
\usage{data(wages)}
\format{
  The format is:
 Time-Series [1:72] from 1982 to 1987: 4.92 4.96 5.04 5.05 5.04 5.04 5.18 5.13 5.15 5.18 ...
}
\source{
Cryer, J. D. Time Series Analysis,
Duxbury Press, 1986.
}
\examples{
data(wages)
## maybe str(wages) ; plot(wages) ...
}
\keyword{datasets}
